Features Usage of Subscription (#4)
* Integrated Devise gem

* Implemented Devise User Model with additional paramters

* Updated README.md file

* Included bootstrap for User Authentication

* Profile Image upload for User completed

* Profile Image upload for User completed

* Featue model added

* Changed User table admin type from string to boolean

* Change column name max_unit_price to max_unit_limit in Features table

* Added index for Features

* Updated .gitignore file

* Removed bin folder from repository

* Removed .ida folder from repository

* Featured CRUD completed

* Fixed format and indentation

* Added extra line at end of files

* Hide Edit/Destory option for simple user

* Plans Crud implemented

* PlansFeatures many-to-many relation created

* Features added with Plan

* Include Features in Plans association completed

* Subscription migration and Model added

* User Subscribe and UnSubscribe a plan functionality completed

* Transaction inserted on Subscribtion

* Subscriptions views completed

* Subscriptions views completed

* Subscriptions views completed

* Transactions view added

* Montly fee calucluation fixed in Edit Plan

* Fixed Plan MontlyFee calculation issue

* Plan show page added

* Plan show page added

* Usage model created

* Updated usage view

* Add usage functionality

* Usage create/update completed

* Charge subscription functionality for admin

* Transaction blocked moved

* Email user on subscription transaction

* Fee calculated based on features usage

* Update schedular for automatic transactions

* Charge subscription functionality added

* Removed BIN files

* Transaction added on charge

* Moved transaction code in subscription

* Moved Billing_Day to helper method

* Fixed extra lines

* Resolved PR #4 comments

* Resolved PR #4 comments on nested routes

* Resolved PR #4 comments Rake task for daily transaction

* Removed exrta code from Subscription controller

* Removed exrta code from Subscription controller

* Fixed emails and urls of example.com

* Resolved PR #4 comments

* Resolved PR #4 comments

* * resolved pr #4 comments

* * resolved pr #4 comments removed if block

+++ db/migrate/20200824072707_create_plans.rb
@@ -2,7 +2,7 @@ class CreatePlans < ActiveRecord::Migration[5.1]
   def change
     create_table :plans do |t|
       t.string :name, null: false
-      t.integer :monthly_fee, null: false, min: 0, default: 0
+      t.float :monthly_fee, null: false, min: 0, default: 0
       t.timestamps
     end
     add_index :plans, :name
